-made changes to bullet pattern codes and also the bullet code themselves to allow changing damage on creating the bullet -Implemented bullet wheel projectile effect for boss -can be placed on scene to see effect. --settings to keep in mind: ---life-time before it expires ---Rotate speed ---activation delay : to allow shooting a distance before activating ---Firing interval: spacing between bullets ---Number of firing points - spots where projectiles are fired from. max 4 ---Projectile force : how fast the projectile travels

+++ TroyPrototype/Assets/Scripts/BossScripts/BossBehaviour.cs
@@ -171,7 +171,7 @@ public class BossBehaviour : MonoBehaviour
     {
         if (collision.gameObject.tag == "ProjectilePlayer")
         {
-            int damageTaken = collision.gameObject.GetComponent<BulletLifetime>().GetDamage();
+            float damageTaken = collision.gameObject.GetComponent<BulletLifetime>().GetDamage();
             if (m_WpnLeft != null)
             {
                 m_WpnLeft.TakeDamage(damageTaken);
@@ -188,7 +188,7 @@ public class BossBehaviour : MonoBehaviour
         }
     }
 
-    public void TakeDamage(int _input)
+    public void TakeDamage(float  _input)
     {
         HP -= _input;
         if(HP<=0)
